@gongguan 2020-07-01
Changed test case because panic will be thrown if NamespaceSelector is nil.
While we can ensure ValidatingWebhook 's NamespaceSelector always not nil, right? Otherwise we can also produce a panic when removing this line https://github.com/kubernetes/kubernetes/blob/master/staging/src/k8s.io/apiserver/pkg/admission/plugin/webhook/generic/webhook_test.go#L114