@gongguan 2020-07-08
Close it because this can't avoid returning an error if the webhook's rules don't even apply to the request.