@gongguan 2020-07-21
address @liggitt https://github.com/kubernetes/kubernetes/pull/92696#discussion_r451012290 and reopen it