@gongguan 2020-07-27
> Exactly. We should not return an error here for webhooks which do not even apply to the incoming request.

@liggitt I have addressed your comment and reopen this pr, PTAL.