@liggitt 2020-07-27
in determining if this is a beneficial change, benchmark-wise, there are a few scenarios we should consider:

* webhook config with complex selector (namespace or object or both) and simple matching rule
* webhook config with simple or absent selector and complex matching rules
* webhook config with complex selector (namespace or object or both) and complex matching rules

Comparing the before/after for all three scenarios will help determine whether there are cases where this change actually decreases performance