@gongguan 2020-07-27
> in determining if this is a beneficial change, benchmark-wise, there are a few scenarios we should consider:

Do you mean I should add at least three benchmark tests to cover these scenarios ?