@liggitt 2020-07-27
> @liggitt Do you mean I should add at least three benchmark tests to cover these scenarios ?

yes, extend the current benchmark to run those three scenarios