@gongguan 2020-07-28
@liggitt I have extended current benchmark tests, the improvement is more obvious:

Before
```
BenchmarkShouldCallHookWithComplexSelector-4              942778              1214 ns/op             288 B/op          6 allocs/op
BenchmarkShouldCallHookWithComplexRule-4                   63638             20492 ns/op              64 B/op          1 allocs/op
BenchmarkShouldCallHookWithComplexSelectorAndRule-4        61670             20800 ns/op              65 B/op          1 allocs/op

```
After
```
BenchmarkShouldCallHookWithComplexSelector-4             6707960               218 ns/op              32 B/op          2 allocs/op
BenchmarkShouldCallHookWithComplexRule-4                 5192454               216 ns/op              32 B/op          2 allocs/op
BenchmarkShouldCallHookWithComplexSelectorAndRule-4      5837241               221 ns/op              32 B/op          2 allocs/op
```