@liggitt 2020-07-28
/lgtm
/approve

it makes sense that a complex selector is faster at short-circuiting than a complex list of rules